+++ src/Plugin/CKEditor5Plugin/SmalkAds.php
@@ -3,8 +3,6 @@
 namespace Drupal\smalk\Plugin\CKEditor5Plugin;
 
 use Drupal\ckeditor5\Plugin\CKEditor5PluginDefault;
-use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * CKEditor5 plugin to allow smalk-ads attribute on div elements.
@@ -27,20 +25,5 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  *   ),
  * )
  */
-class SmalkAds extends CKEditor5PluginDefault implements ContainerFactoryPluginInterface {
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition) {
-    return new static(
-      $configuration,
-      $plugin_id,
-      $plugin_definition
-    );
-  }
-
+class SmalkAds extends CKEditor5PluginDefault {
 }
-
-
-